# src/api/api.yaml
openapi: 3.0.0
paths:
  /person/list:
    get:
      operationId: PersonController_listPersons
      parameters:
        - name: take
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 20
            type: number
        - name: skip
          required: false
          in: query
          schema:
            minimum: 0
            default: 0
            type: number
      responses:
        '200':
          description: List all persons for the current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonListResponseDto'
      tags: &ref_0
        - Person
      security: &ref_1
        - bearer: []
  /person/{id}:
    get:
      operationId: PersonController_getPerson
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Get person details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDetailResponseDto'
      tags: *ref_0
      security: *ref_1
    put:
      operationId: PersonController_updatePerson
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonUpdateDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDetailResponseDto'
      tags: *ref_0
      security: *ref_1
    delete:
      operationId: PersonController_deletePerson
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDetailResponseDto'
      tags: *ref_0
      security: *ref_1
  /person/{id}/ai/history:
    get:
      operationId: PersonController_getAiHistory
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiHistoryListResponseDto'
      tags: *ref_0
      security: *ref_1
  /person:
    post:
      operationId: PersonController_createPerson
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonCreateDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDetailResponseDto'
      tags: *ref_0
      security: *ref_1
...
  /person/ai/{id}:
    get:
      operationId: PersonAiController_getPersonForAI
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: 'AI: Get person details (with AI context)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonAiDetailResponseDto'
      tags: *ref_15
    put:
      operationId: PersonAiController_updatePerson
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonAiUpdateDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonAiDetailResponseDto'
      tags: *ref_15
  /person/ai:
    post:
      operationId: PersonAiController_createPerson
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonAiCreateDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonAiDetailResponseDto'
      tags: *ref_15
  /person/ai/{id}/history:
    get:
      operationId: PersonAiController_getAiHistory
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiHistoryListResponseDto'
      tags: *ref_15
  /person/ai/{id}/context:
    get:
      operationId: PersonAiController_getAiContext
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiContextDto'
      tags: *ref_15
    put:
      operationId: PersonAiController_updateAiContext
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiContextDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiContextDto'
      tags: *ref_15
  /place/list:
    get:
      operationId: PlaceController_listPlaces
      parameters:
        - name: take
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 20
            type: number
        - name: skip
          required: false
          in: query
          schema:
            minimum: 0
            default: 0
            type: number
      responses:
        '200':
          description: List all places for the current user (owner or member)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceListResponseDto'
      tags: &ref_16
        - Place
      security: &ref_17
        - bearer: []
  /place/{id}:
    get:
      operationId: PlaceController_getPlace
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Get place details (with members and location)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceDetailResponseDto'
      tags: *ref_16
      security: *ref_17
    put:
      operationId: PlaceController_updatePlace
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceUpdateDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceDetailResponseDto'
      tags: *ref_16
      security: *ref_17
    delete:
      operationId: PlaceController_deletePlace
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceDetailResponseDto'
      tags: *ref_16
      security: *ref_17
...
  /place/ai/{id}:
    get:
      operationId: PlaceAiController_getPlace
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceAiDetailResponseDto'
      tags: *ref_18
    put:
      operationId: PlaceAiController_updatePlace
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceAiUpdateDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceAiDetailResponseDto'
      tags: *ref_18
  /place/ai:
    post:
      operationId: PlaceAiController_createPlace
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceAiCreateDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceAiDetailResponseDto'
      tags: *ref_18
  /place/ai/{id}/history:
    get:
      operationId: PlaceAiController_getAiHistory
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiHistoryListResponseDto'
      tags: *ref_18
  /place/ai/{id}/context:
    get:
      operationId: PlaceAiController_getAiContext
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiContextResponseDto'
      tags: *ref_18
    put:
      operationId: PlaceAiController_updateAiContext
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiContextDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiContextResponseDto'
      tags: *ref_18
...
info:
  title: WHCP Backend
  description: XD
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
    JWT-refresh:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT refresh
      description: Enter JWT refresh token
      in: header
    JWT-reset-password:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT reset password
      description: Enter JWT reset password token
      in: header
    JWT-register-user:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT register user
      description: Enter JWT register user
      in: header
  schemas:
    PersonRelation:
      type: string
      enum:
        - FAMILY
        - FRIEND
        - WORK
        - OTHER
    AiContextResponseDto:
      type: object
      properties:
        id:
          type: string
        summary:
          type: string
        keywords:
          type: object
        context:
          type: string
        aiMeta:
          type: object
        aiConfidence:
          type: number
        aiSource:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
    PersonDetailResponseDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nickname:
          type: string
        email:
          type: string
        phone:
          type: string
        notes:
          type: string
        birthday:
          type: string
          format: date-time
        avatarUrl:
          type: string
        relation:
          $ref: '#/components/schemas/PersonRelation'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        aiContext:
          $ref: '#/components/schemas/AiContextResponseDto'
      required:
        - id
        - name
        - createdAt
        - updatedAt
    PaginationDto:
      type: object
      properties:
        take:
          type: number
          minimum: 1
          maximum: 20
        skip:
          type: number
          minimum: 0
          default: 0
    PersonListResponseDto:
      type: object
      properties:
        persons:
          type: array
          items:
            $ref: '#/components/schemas/PersonDetailResponseDto'
        total:
          type: number
        params:
          $ref: '#/components/schemas/PaginationDto'
      required:
        - persons
        - total
    AiHistoryResponseDto:
      type: object
      properties:
        id:
          type: string
        aiContextId:
          type: string
        timestamp:
          type: string
          format: date-time
        action:
          type: string
        source:
          type: string
        userPrompt:
          type: string
        aiResponse:
          type: string
        text:
          type: string
        meta:
          type: object
      required:
        - id
        - aiContextId
        - timestamp
        - action
    AiHistoryListResponseDto:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/AiHistoryResponseDto'
        total:
          type: number
      required:
        - entries
        - total
    PersonCreateDto:
      type: object
      properties:
        name:
          type: string
        nickname:
          type: string
        email:
          type: string
        phone:
          type: string
        notes:
          type: string
        birthday:
          type: string
          format: date-time
        avatarUrl:
          type: string
        relation:
          $ref: '#/components/schemas/PersonRelation'
        locationId:
          type: string
          description: Location ID to associate with this person
      required:
        - name
    PersonUpdateDto:
      type: object
      properties:
        name:
          type: string
        nickname:
          type: string
        email:
          type: string
        phone:
          type: string
        notes:
          type: string
        birthday:
          type: string
          format: date-time
        avatarUrl:
          type: string
        relation:
          $ref: '#/components/schemas/PersonRelation'
        locationId:
          type: string
          description: Location ID to associate with this person
...
    AiContextDto:
      type: object
      properties:
        userPrompt:
          type: string
          description: Prompt from user to the AI
        aiResponse:
          type: string
          description: AI-generated response
        aiMeta:
          type: object
          description: Meta/object for AI log details
        summary:
          type: string
        keywords:
          type: object
        context:
          type: string
        aiConfidence:
          type: number
        aiSource:
          type: string
    PersonAiDetailResponseDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nickname:
          type: string
        email:
          type: string
        phone:
          type: string
        notes:
          type: string
        avatarUrl:
          type: string
        relation:
          $ref: '#/components/schemas/PersonRelation'
        aiContext:
          $ref: '#/components/schemas/AiContextDto'
      required:
        - id
        - name
...
    PersonAiCreateDto:
      type: object
      properties:
        userPrompt:
          type: string
          description: Prompt from user to the AI
        aiResponse:
          type: string
          description: AI-generated response
        aiMeta:
          type: object
          description: Meta/object for AI log details
        name:
          type: string
        nickname:
          type: string
        email:
          type: string
        phone:
          type: string
        notes:
          type: string
        birthday:
          type: string
          format: date-time
        relation:
          $ref: '#/components/schemas/PersonRelation'
      required:
        - name
    PersonAiUpdateDto:
      type: object
      properties:
        userPrompt:
          type: string
          description: Prompt from user to the AI
        aiResponse:
          type: string
          description: AI-generated response
        aiMeta:
          type: object
          description: Meta/object for AI log details
        name:
          type: string
        nickname:
          type: string
        email:
          type: string
        phone:
          type: string
        notes:
          type: string
        birthday:
          type: string
          format: date-time
        relation:
          $ref: '#/components/schemas/PersonRelation'
    PlaceType:
      type: string
      enum:
        - HOME
        - WORK
        - FAMILY_HOME
        - FRIEND_HOME
        - RESTAURANT
        - SHOP
        - OFFICE
        - SCHOOL
        - GYM
        - HOSPITAL
        - HOTEL
        - HOSTEL
        - AIRPORT
        - TRAIN_STATION
        - BUS_STATION
        - PORT
        - PARK
        - BEACH
        - MOUNTAIN
        - CAMP
        - NATIONAL_PARK
        - MUSEUM
        - MONUMENT
        - ATTRACTION
        - TRAVEL_DESTINATION
        - SIGHTSEEING
        - EVENT
        - OTHER
    LocationResponseDto:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
        radius:
          type: integer
          description: Radius in meters
          example: 50
        latitude:
          type: number
        longitude:
          type: number
        description:
          type: string
    PlaceDetailResponseDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          $ref: '#/components/schemas/PlaceType'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        aiContext:
          $ref: '#/components/schemas/AiContextResponseDto'
        location:
          $ref: '#/components/schemas/LocationResponseDto'
      required:
        - id
        - name
    PlaceListResponseDto:
      type: object
      properties:
        places:
          type: array
          items:
            $ref: '#/components/schemas/PlaceDetailResponseDto'
        total:
          type: number
        params:
          $ref: '#/components/schemas/PaginationDto'
      required:
        - places
        - total
    LocationCreateDto:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
        radius:
          type: integer
          description: Radius in meters
          example: 50
        latitude:
          type: number
        longitude:
          type: number
        description:
          type: string
...
    PlaceUpdateDto:
      type: object
      properties:
        userPrompt:
          type: string
          description: Prompt from user to the AI
        aiResponse:
          type: string
          description: AI-generated response
        aiMeta:
          type: object
          description: Meta/object for AI log details
        name:
          type: string
        description:
          type: string
        type:
          $ref: '#/components/schemas/PlaceType'
        location:
          $ref: '#/components/schemas/LocationCreateDto'
...
    PlaceAiDetailResponseDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          $ref: '#/components/schemas/PlaceType'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        aiContext:
          $ref: '#/components/schemas/AiContextResponseDto'
        location:
          $ref: '#/components/schemas/LocationResponseDto'
      required:
        - id
        - name
...
    PlaceAiCreateDto:
      type: object
      properties:
        userPrompt:
          type: string
          description: Prompt from user to the AI
        aiResponse:
          type: string
          description: AI-generated response
        aiMeta:
          type: object
          description: Meta/object for AI log details
        name:
          type: string
        description:
          type: string
        type:
          $ref: '#/components/schemas/PlaceType'
        location:
          $ref: '#/components/schemas/LocationCreateDto'
      required:
        - name
    PlaceAiUpdateDto:
      type: object
      properties:
        userPrompt:
          type: string
          description: Prompt from user to the AI
        aiResponse:
          type: string
          description: AI-generated response
        aiMeta:
          type: object
          description: Meta/object for AI log details
        name:
          type: string
        description:
          type: string
        type:
          $ref: '#/components/schemas/PlaceType'
        location:
          $ref: '#/components/schemas/LocationCreateDto'
...